+++ packages/ua-devtools-evm-hardhat/src/ownable/types.ts
@@ -1,0 +1,4 @@
+import type { OmniGraphHardhat } from '@layerzerolabs/devtools-evm-hardhat'
+import type { OwnableNodeConfig } from '@layerzerolabs/ua-devtools'
+
+export type OwnableOmniGraphHardhat = OmniGraphHardhat<OwnableNodeConfig>